:hammer: [Update] 현재 활성화 탭 이동 및 함수 이름 변경 #main
