+++ src/router.js
@@ -3,7 +3,21 @@ import HomePage from "./pages/Home.js";
 import ProfilePage from "./pages/Profile.js";
 import NotFoundPage from "./pages/NotFound.js";
 import Layout from "./layout/Layout.js";
-import { handleTabClick } from "./eventHandlers.js";
+
+const updateActiveTabFromURL = () => {
+  const tabs = document.querySelectorAll("a.tab");
+  tabs.forEach((tab) => {
+    tab.classList.remove("text-blue-600");
+    tab.classList.remove("font-bold");
+    tab.classList.add("text-gray-600");
+  });
+  const currentTab = window.location.pathname;
+  const tab = document.querySelector(`a[href="${currentTab}"]`);
+  if (!tab) return;
+  tab.classList.remove("text-gray-600");
+  tab.classList.add("text-blue-600");
+  tab.classList.add("font-bold");
+};
 
 const renderPage = (path) => {
   const currentRoute = path || window.location.pathname;
@@ -34,7 +48,7 @@ const renderPage = (path) => {
   }
 
   document.querySelector("#root").innerHTML = currentPage;
-  handleTabClick();
+  updateActiveTabFromURL();
 };
 
 const navigate = (path) => {
